Inline training loop into trainer
